feat: factor out task runs hook
